requirements.txt
halo
mysql-connector
matplotlib
bs4
xmltodict
Naked
habanero
networkx
pandas
webweb
metaknowledge
jupyter
refextract
gensim
pyLDAvis
nltk
spacy
surprise
google-api-python-client
fuzzywuzzy
openpyxl